+++ index.ts
@@ -329,7 +329,7 @@ import { initFilterBackend } from './src/filters/FilterBackend';
 import { Canvas2dFilterBackend } from './src/filters/2d_backend.class';
 import { WebGLFilterBackend } from './src/filters/webgl_backend.class';
 import { runningAnimations } from './src/util/animation/AnimationRegistry';
-import { Observable } from './src/mixins/observable.mixin';
+import { Observable } from './src/mixins/Observable';
 import { Point } from './src/point.class';
 import { Intersection } from './src/intersection.class';
 import { Color } from './src/color/color.class';
@@ -358,7 +358,7 @@ import { Group } from './src/shapes/group.class';
 import { ActiveSelection } from './src/shapes/active_selection.class';
 import { Image } from './src/shapes/image.class';
 import { getEnv, getDocument, getWindow, setEnvForTests } from './src/env';
-import { createCollectionMixin } from './src/mixins/collection.mixin';
+import { createCollectionMixin } from './src/mixins/Collection';
 import { parseAttributes } from './src/parser/parseAttributes';
 import { parseElements } from './src/parser/parseElements';
 import { getFilterBackend } from './src/filters/FilterBackend';
