+++ src/canvas/static_canvas.class.ts
@@ -3,8 +3,8 @@ import { config } from '../config';
 import { iMatrix, VERSION } from '../constants';
 import type { CanvasEvents, StaticCanvasEvents } from '../EventTypeDefs';
 import type { Gradient } from '../gradient/gradient.class';
-import { createCollectionMixin } from '../mixins/collection.mixin';
-import { CommonMethods } from '../mixins/shared_methods.mixin';
+import { createCollectionMixin } from '../mixins/Collection';
+import { CommonMethods } from '../mixins/CommonMethods';
 import type { Pattern } from '../pattern.class';
 import { Point } from '../point.class';
 import type { BaseFabricObject as FabricObject } from '../EventTypeDefs';
